requirements.txt
python-dotenv
nnsight==0.3.7
matplotlib
seaborn
plotly
wandb 
tqdm
jaxtyping
einops
git+https://github.com/dsbowen/strong_reject.git@main 
jupyterlab
accelerate==1.2.0
einops==0.8.0
torch==2.5.1
transformers==4.47.0
